feat: The structure and logic of the package was updated.
In this commit, the package comes with more modular codes and also with new features. In this feature, you have to select the content of the logger as {}, which has 4 options, 2 of which are mandatory, namely message and format. Also, the default level is info. And you don't need to specify it every time (unless you want something else).

A new namespace called Argument was added, which contains the type of input arguments and all arguments are written in it.

A new option is received as an input that you can specify the color of your logs in the console. Personalize.